requirements.txt
streamlit
psycopg2-binary
python-dotenv
langchain
langchain-openai
langgraph
rapidfuzz
pandas
